Split Dashboard and Coverage into dedicated controllers
Each surface now has its own controller and template directory:

- Admin\DashboardController::index — was AuditLogsController::dashboard
- Admin\CoverageController::index — was AuditLogsController::coverage
- AuditLogsController stays focused on browsing/inspecting logs

Shared admin init (layout setup + AuditStash.accessCheck gate +
plugin helper loading) extracts to Admin\AdminControllerTrait, used
by all three controllers. Trims the controller bodies and removes
the duplication that would otherwise grow with each new surface.

Routing collapses to the minimum: plugin root → Dashboard::index,
extensions for csv/json export, fallbacks. The previous explicit
/view, /timeline, /export, /audit-logs/<...> connects are gone —
fallbacks serves them at /audit-logs/<action> with the same set of
parameters.

config/routes.php is now an explicit no-op stub. The plugin's routes
live in AuditStashPlugin::routes(); the file exists only to satisfy
BaseApplication::routes() which require()s it during host-app boot.
